[clean] Fix comparison warning w/ 0 effect params
This is a rather crude fix for the warning, but it's probably the
simplest one. If an effect has zero parameters, the compiler
(correctly) complains that the comparison in the for loop condition
is always false.

However we don't want to have a) another specific macro for this case,
or b) check this case explicitly (since the for loop does exactly the
same thing). Instead we make the index variable signed (which silences
the warning) and larger (to avoid overflow, however unlikely).

+++ avs/vis_avs/effect_info.h
@@ -635,7 +635,7 @@ struct Effect_Info {
     virtual const Parameter* get_parameters() const { return this->parameters; }; \
     virtual const AVS_Parameter_Handle* get_parameters_for_api() const {          \
         static AVS_Parameter_Handle parameters_for_api[this->num_parameters];     \
-        for (uint32_t i = 0; i < this->num_parameters; i++) {                     \
+        for (int64_t i = 0; i < this->num_parameters; i++) {                      \
             parameters_for_api[i] = this->parameters[i].handle;                   \
         }                                                                         \
         return parameters_for_api;                                                \
